requirements.txt
fastapi
uvicorn[standard]
aiosqlite
aiohttp
google-generativeai
pydantic
anyio>=4.8.0